refactor(backup): move backup logic to internal core layer (#12)
