+++ internal/core/backup/dotfile_scanner.go
